adsServiceImpl methods

+++ src/main/java/com/skyteam/shopping_area/dto/AdDto.java
@@ -23,9 +23,10 @@ public class AdDto {
     @Schema(description = "ссылка на картинку объявления")
     private String image;
     @Schema(description = "id объявления")
-    private int pk;
+    private long pk;
     @Schema(description = "цена объявления")
     private int price;
     @Schema(description = "заголовок объявления")
     private String title;
+
 }
